requirements.txt
azureml-sdk
azure-blob-storage
azure-identity
azure-keyvault

-e my_custom_package